added dockerfile and start test coverage

+++ src/main/java/ru/pranch/cinema/services/CinemaService.java
@@ -121,7 +121,8 @@ public class CinemaService {
         }
       });
 
-    cinemaHallDao.deleteAllById(cinemaHallsIds);
+    if (!cinemaHallsIds.isEmpty())
+      cinemaHallDao.deleteAllById(cinemaHallsIds);
 
     return cinemaDao.deleteById(cinemaId);
   }
